refactor: move audit and soft_delete into models subpackage
Group model-enhancement modules (mixins, helpers, directives) under
pyramid_sa/models/ for clearer package structure. Consolidate the
duplicate _now() into a shared utils.py alongside generate_uuid().

The top-level public API (from pyramid_sa import ...) is unchanged.

Closes #25
