+++ docs/models.md
@@ -57,7 +57,7 @@ Example with selective mixins:
 
 ```python
 from pyramid_sa import Base
-from pyramid_sa.audit import AuditMixin
+from pyramid_sa.models.audit import AuditMixin
 
 
 class LogEntry(AuditMixin, Base):
